# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PokeAPI",
        "contact": {
            "name": "Komang Damai",
            "email": "komangdamai3@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "gopoke-production.up.railway.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/pokemon": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of pokemons by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemons"
                ],
                "summary": "Get pokemon by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden access",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Pokemon not found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    }
                }
            }
        },
        "/pokemon/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creating new pokemon with role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemons"
                ],
                "summary": "Create new pokemon",
                "parameters": [
                    {
                        "description": "Create pokemon credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFavoritePokemonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Token"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden access",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "Pokemon email already used",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to hash password",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    }
                }
            }
        },
        "/pokemon/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete pokemon by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemons"
                ],
                "summary": "Delete pokemon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Token"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden access",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Pokemons not found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    }
                }
            }
        },
        "/pokemon/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update pokemon by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemons"
                ],
                "summary": "Update pokemon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Update pokemon credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFavoritePokemonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Token"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden access",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Pokemons not found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    }
                }
            }
        },
        "/pokemons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of pokemons with filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pokemons"
                ],
                "summary": "Get all pokemons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by notes",
                        "name": "notes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field (name, type, notes)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc or desc), default is asc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden access",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Pokemons not found",
                        "schema": {
                            "$ref": "#/definitions/utils.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateFavoritePokemonRequest": {
            "type": "object",
            "required": [
                "notes"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 30
                },
                "sprite": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.TimeJSON": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "dto.Token": {
            "type": "object",
            "properties": {
                "activeUntil": {
                    "$ref": "#/definitions/dto.TimeJSON"
                },
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateFavoritePokemonRequest": {
            "type": "object",
            "required": [
                "notes"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 30
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utils.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}